Finally put to rest the debate about _free() signatures
Previously, many of the libraries free functions took **, in order
to be able to set the original pointer to NULL, and (assuming the
same free function did a check for this at the beginning of the
call) prevent double-frees. I decided it was much more valuable to
have free functions with signatures congruent to free(3).